working super fast version of angular coefficients. physics to be checked
