+++ .gitignore
@@ -4,7 +4,3 @@ __init__.py
 .#*
 #*#
 *~
-build
-*.d
-*.so
-*.pcm
