--- framework/.gitignore
@@ -1,6 +1,0 @@
-__init__.py
-*.pyc
-.*.swp
-.#*
-#*#
-*~
