+++ wmass/AngCoeff.h
@@ -1,0 +1,41 @@
+#ifndef ANGCOEFF_H
+#define ANGCOEFF_H
+
+
+#include "ROOT/RDataFrame.hxx"
+#include "ROOT/RVec.hxx"
+#include "ROOT/RDF/RInterface.hxx"
+#include "TH1D.h"
+#include "TH2D.h"
+#include "TString.h"
+#include "TMath.h"
+#include "../framework/module.h"
+
+using namespace ROOT::VecOps;
+using RNode = ROOT::RDF::RNode;
+using rvec_f = const RVec<float> &;
+using rvec_i = const RVec<int> &;
+
+
+class AngCoeff : public Module {
+
+    private:
+
+    std::vector<ROOT::RDF::RResultPtr<TH1D>> _h1List;
+    std::vector<ROOT::RDF::RResultPtr<TH2D>> _h2List;
+    
+    public:
+    
+    AngCoeff() {};
+    ~AngCoeff() {};
+    RNode doSomething(RNode) override;
+    std::vector<ROOT::RDF::RResultPtr<TH1D>> getTH1() override;
+    std::vector<ROOT::RDF::RResultPtr<TH2D>> getTH2() override;
+
+};
+
+#endif
+
+
+
+
